moved cli utils
